HBASE-12444 Total number of requests overflow because it's int

+++ hbase-rest/src/main/java/org/apache/hadoop/hbase/rest/model/StorageClusterStatusModel.java
@@ -630,7 +630,7 @@ public class StorageClusterStatusModel
    * @param requests the total number of requests per second handled by the
    * cluster
    */
-  public void setRequests(int requests) {
+  public void setRequests(long requests) {
     this.requests = requests;
   }
